temp commit

+++ src/main/java/com/neodymium/davisbase/models/CellHeader.java
@@ -5,7 +5,7 @@ public interface CellHeader {
 
     short leftChildPage();
 
-    short size();
+    byte size();
 
     int rowId();
 }
